pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs", "requests"]
build-backend = "hatchling.build"


[project]
name = "panda-server"
dynamic = ["version"]
description = "PanDA Server Package"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    { name = "PanDA Team", email = "panda-support@cern.ch" },
]
dependencies = [
    'panda-common>=0.1.5',
    'panda-client-light>=1.5.55',
    'pyOpenSSL',
    'python-daemon',
    'mod_wsgi',
    'stomp.py',
    'pyyaml',
    'pyjwt',
    'polars',
    'requests',
    'psutil>=5.4.8',
    'idds-common',
    'idds-client',
    'idds-workflow>=1.1.0',
    'idds-doma',
    'idds-atlas',
    'ruamel.yaml',
    'cwl-utils>=0.13',
    'packaging',
    'snakemake==7.30.1',
    'numpy',
    'scipy',
    'werkzeug',
    'requests',
]
requires-python = ">=3.10"

[project.optional-dependencies]
oracle = ['oracledb']
mysql = ['mysqlclient']
postgres =  ['psycopg2-binary']
rucio = ['rucio-clients']
elasticsearch = ['elasticsearch']
atlasprod =  ['oracledb', 'rucio-clients', 'elasticsearch', 'numpy', 'scipy']
mcp = ['uvicorn', 'fastmcp']

[project.urls]
Homepage = "https://panda-wms.readthedocs.io/en/latest/"

[tool.hatch.version]
path = "PandaPkgInfo.py"
pattern = "release_version = \"(?P<version>[^\"]+)\""

[tool.hatch.build]
directory = "dist"

[tool.hatch.build.targets.wheel]
exclude = ["*.template"]
packages = ["pandaserver", "package"]

[tool.hatch.build.targets.wheel.shared-data]
"templates" = "etc/panda"
"templates/bin" = "usr/bin"
"templates/init.d" = "etc/rc.d/init.d"
"templates/systemd" = "etc/systemd/system"
"templates/sysconfig" = "etc/panda"

[tool.hatch.build.targets.wheel.hooks.custom]
path = "package/hatch_build.py"

[tool.hatch.build.targets.sdist]
exclude = [
    ".github",
    ".idea",
]

[tool.black]
line-length=160

[tool.autopep8]
# https://pypi.org/project/autopep8/#pyproject-toml
max_line_length = 160
ignore = ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3

[tool.pylint.master]
max-line-length = 160
[tool.pylint.'MESSAGES CONTROL']
disable = [
    "C0209",
    "R0902",
    "R0913",
    "R0914", #  R0914: Too many local variables
]

[tool.isort]
profile = "black"

[tool.flynt]
line-length = 160

[tool.flake8]
ignore = ['E501', 'U100']